Optimize radius type to be Float

+++ AAInfographics/AAChartCreator/AAChartModel.swift
@@ -188,8 +188,8 @@ public class AAChartModel: AAObject {
     public var series: [AASeriesElement]?   //An array of all the chart's series
     public var legendEnabled: Bool?         //Enable or disable the legend. Defaults to true
     public var backgroundColor: Any?        //The background color or gradient for the outer chart area. Defaults to #FFFFFF
-    public var borderRadius: Int?           //The corner radius of the outer chart border. Defaults to 0
-    public var markerRadius: Int?           //The radius of the point marker. Defaults to 4
+    public var borderRadius: Float?           //The corner radius of the outer chart border. Defaults to 0
+    public var markerRadius: Float?           //The radius of the point marker. Defaults to 4
     public var touchEventEnabled: Bool?     //Support touch event call back or not
     
     @discardableResult
@@ -277,7 +277,7 @@ public class AAChartModel: AAObject {
     }
     
     @discardableResult
-    public func markerRadius(_ prop: Int) -> AAChartModel {
+    public func markerRadius(_ prop: Float) -> AAChartModel {
         markerRadius = prop
         return self
     }
@@ -481,7 +481,7 @@ public class AAChartModel: AAObject {
     }
     
     @discardableResult
-    public func borderRadius(_ prop: Int) -> AAChartModel {
+    public func borderRadius(_ prop: Float) -> AAChartModel {
         borderRadius = prop
         return self
     }
